sourceCode/requirements.txt
# Base requirements
numpy==1.23.5
opencv-python==4.7.0.72
torch==1.13.1
Pillow==9.4.0
pyyaml==6.0
requests==2.28.1

# PaddleOCR and PaddlePaddle
paddlepaddle==2.4.2
paddleocr==2.6.1.3

# Additional requirements
matplotlib==3.6.2
scipy==1.9.3
